Implemented Template for Stores
